+++ pkg/config/resources/exec/resource.go
@@ -33,7 +33,9 @@ type Exec struct {
 	RunAs    *ctypes.User               `hcl:"run_as,block" json:"run_as,omitempty"`    // User block for mapping the user id and group id inside the container
 
 	// output
-	PID int `hcl:"pid,optional" json:"pid,omitempty"` // PID stores the ID of the created connector service if it is a local exec
+	PID      int               `hcl:"pid,optional" json:"pid,omitempty"`             // PID stores the ID of the created connector service if it is a local exec
+	ExitCode int               `hcl:"exit_code,optional" json:"exit_code,omitempty"` // Exit code of the process
+	Output   map[string]string `hcl:"output,optional" json:"output,omitempty"`       // output values returned from exec
 }
 
 func (e *Exec) Process() error {
@@ -63,6 +65,8 @@ func (e *Exec) Process() error {
 		if r != nil {
 			kstate := r.(*Exec)
 			e.PID = kstate.PID
+			e.ExitCode = kstate.ExitCode
+			e.Output = kstate.Output
 		}
 	}