+++ ChangeLog.md
@@ -1,8 +1,54 @@
 # Change Log
 
-## version v0.10.1
-* Ensure that the total compute is set correctly	for Nomad clusters when 
-	running on Docker in Apple Silicon.
+## version v0.10.5
+
+## New Features:
+Adds outputs for `exec` resources
+
+Exec resources now have a new parameter `output` which is a map of key value pairs.
+Values for output can be set by echoing a key value to the file `${EXEC_OUTPUT}` in 
+the defined script for either remote or local exec.
+
+```hcl
+resource "exec" "install" {
+  # Add the output
+  echo "exec=install" >> $EXEC_OUTPUT
+  echo "foo=bar" >> $EXEC_OUTPUT
+  EOF
+
+  timeout = "30s"
+}
+
+output "local_exec_install" {
+  value = resource.exec.install.output.exec
+}
+```
+
+## version v0.10.4
+
+## New Features:
+Enable experimental support for nvidia GPUs for container resources
+
+This feature configures the container to use the nvidia runtime and the nvidia
+device plugin to access the GPU.  Currently this has only been tested with WSL2 and
+Nvidia GPUs.
+
+```hcl
+resource "container" "gpu_test" {
+  image {
+    name = "nvcr.io/nvidia/k8s/cuda-sample:nbody"
+  }
+
+  command = ["nbody", "-gpu", "-benchmark"]
+
+  resources {
+    gpu {
+      driver     = "nvidia"
+      device_ids = ["0"]
+    }
+  }
+}
+```
 
 ## version v0.10.0